Rename function to avoid conflicting with `process` global
