+++ .github/actions/high-priority-prs/src/__tests__/pr-message.js
@@ -1,4 +1,4 @@
-const { process, maintainers } = require("../process")
+const { processData, maintainers } = require("../process-data")
 const report = require("../pr-message")
 const fs = require("fs")
 const path = require("path")
@@ -8,7 +8,7 @@ const fixture = JSON.parse(
 
 test("Creates a correctly formatted Slack message", () => {
   const now = new Date(`2019-06-05T13:49:58Z`)
-  const queues = process(fixture, now)
+  const queues = processData(fixture, now)
   const slackMessage = report(queues, maintainers, now)
   expect(slackMessage).toMatchSnapshot()
