+++ .github/actions/high-priority-prs/src/__tests__/process-data.js
@@ -1,4 +1,4 @@
-const { process } = require("../process")
+const { processData } = require("../process-data")
 const fs = require("fs")
 const path = require("path")
 const fixture = JSON.parse(
@@ -7,6 +7,8 @@ const fixture = JSON.parse(
 
 test("Gets PRs that have not been updated in at least 30 days, sorted by oldest first", () => {
   const now = new Date(`2019-06-05T13:49:58Z`)
-  const lonelyPrDates = process(fixture, now).lonelyPrs.map(pr => pr.updatedAt)
+  const lonelyPrDates = processData(fixture, now).lonelyPrs.map(
+    pr => pr.updatedAt
+  )
   expect(lonelyPrDates).toMatchSnapshot()
 })
