+++ .github/actions/high-priority-prs/src/index.js
@@ -1,9 +1,9 @@
 const fetch = require("./fetch.js")
-const { process, report } = require("./process.js")
+const { processData, report } = require("./process-data.js")
 
 const start = async function() {
   const data = await fetch()
-  if (data) report(process(data))
+  if (data) report(processData(data))
 }
 
 start()
